requirements.txt
discord.py
python-dotenv
aiosmtplib
beanie
motor